requirements.txt
fastapi==0.103.2
aiogram==3.1.1
sqlalchemy==2.0.22
sqlalchemy-utils==0.41.1
loguru==0.7.2
pyjwt==2.8.0
pydantic-settings==2.0.3
asyncpg==0.28.0 
psycopg2-binary==2.9.9
uvicorn==0.23.2
pydantic[email]
pytz==2023.3.post1
stackprinter==0.2.10
passlib==1.7.4
greenlet==3.0.0
bcrypt==4.0.1
